Restructure base folder setup
