+++ client/pages/_document.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import Document, { Html, Head, Main, NextScript } from "next/document";
 import createEmotionServer from "@emotion/server/create-instance";
-import theme from "../src/theme";
-import createEmotionCache from "../src/createEmotionCache";
+import theme from "../src/utils/theme";
+import createEmotionCache from "../src/utils/createEmotionCache";
 
 export default class MyDocument extends Document {
   render() {
