+++ client/src/components/SearchBox/index.tsx
@@ -1,0 +1,9 @@
+import useStyles from "./styles";
+
+const index = () => {
+  const classes = useStyles();
+
+  return <div>index</div>;
+};
+
+export default index;
