+++ client/src/components/SearchBox/styles.ts
@@ -1,0 +1,6 @@
+import { Theme } from "@mui/material";
+import { makeStyles } from "@mui/styles";
+
+export default makeStyles((theme: Theme) => ({
+  container: {},
+}));
